Simplify reading of the frequency field from nav.dat files
The frequency field is always an integer, so reading it as a float to
convert it to int doesn't make sense. I've probably been confused by
field indices when I introduced this in commit a2bf4241183a.

This clarifies things, but shouldn't change the code behavior in any
way (unless with bogus nav.dat files, of course).