Experimental multi threaded rendering
The basic theory for this is that after the update traversal there should be no access from the scene graph (OSG) to the property tree and therefore by splitting the rendering into discrete steps (rather than using get_renderer()->update() we could have a second thread for the simulation that would be released after the update traversal and therefore run concurrently with the drawing on the GPU.

To achieve this we have to run the simulation loop after the rendering (previously it was always run before) - so this changes the latency by a small amount.

Small changes had to be made to the scenery loader to ensure that the rendering and loading threads cooperate better. This is the result of testing showing an occasional exception in the loader that appeared to be related to the update loop still being in progress.

Testing shows

1. Reasonable reliability
2. Something causes the aircraft position to occasionally move - as though the position from the next frame was being used.

 investigation and validation it appears that after the update traversal  during the
This is an experimental development that permits the s